refactored version of the code, mostly of main.py
